refactor: @lib 폴더 코어화
