+++ src/@lib/index.ts
@@ -1,3 +1,3 @@
 export * from "./equalities";
-export * from "./hooks";
 export * from "./hocs";
+export * from "./hooks";
