+++ src/providers/index.tsx
@@ -2,7 +2,9 @@ import { NotificationProvider } from "./notification-provider";
 import { ThemeProvider } from "./theme-provider";
 import { UserProvider } from "./user-provider";
 
-const Providers: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+export const Providers: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
   return (
     <NotificationProvider>
       <UserProvider>
@@ -11,5 +13,3 @@ const Providers: React.FC<{ children: React.ReactNode }> = ({ children }) => {
     </NotificationProvider>
   );
 };
-
-export default Providers;
